requirements.txt
PyYAML~=6.0.1
requests~=2.32.3
clang~=17.0.6
nbformat~=5.10.4